// apps/main/src/common/infrastructure/database/types/db.d.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface WsAdditionalNeeds {
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  id: Generated<number>;
  material_need_id: number;
  material_target_id: number;
  remaining_stock: number | null;
  total: number | null;
  updated_at: Generated<Date>;
  updated_by: number | null;
}
// ...
export interface WsMaterialNeeds {
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  id: Generated<number>;
  material_target_id: number;
  microplanning_id: number;
  reference_id: number;
  reference_type: string;
  status: Generated<number | null>;
  total_needs: number | null;
  updated_at: Generated<Date>;
  updated_by: number | null;
}

export interface WsMaterialNeedsDetails {
  absolute_number_of_routine_immunization: number | null;
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  id: Generated<number>;
  material_need_id: number;
  number_of_vials_used: number | null;
  remaining_stock: number | null;
  updated_at: Generated<Date>;
  updated_by: number | null;
}
// ...
export interface WsMonthlyVaccineNeedDetails {
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  id: Generated<number>;
  material_need_id: number;
  max_stock: number | null;
  min_stock: number | null;
  request_qty: number | null;
  updated_at: Generated<Date>;
  updated_by: number | null;
}
// ...
export interface WsVaccineUtilizationRate {
  created_at: Generated<Date>;
  created_by: number | null;
  deleted_at: Date | null;
  deleted_by: number | null;
  id: Generated<number>;
  material_need_id: number;
  updated_at: Generated<Date>;
  updated_by: number | null;
  vaccine_utilization_rate: number | null;
}
// ...
export interface DB {
// ...
  ws_additional_needs: WsAdditionalNeeds;
// ...
  ws_material_needs: WsMaterialNeeds;
  ws_material_needs_details: WsMaterialNeedsDetails;
  ws_material_permissions: WsMaterialPermissions;
  ws_material_ratios: WsMaterialRatios;
  ws_material_substitutions: WsMaterialSubstitutions;
  ws_material_targets: WsMaterialTargets;
  ws_materials: WsMaterials;
  ws_microplan_targets_consumptions: WsMicroplanTargetsConsumptions;
  ws_microplanning: WsMicroplanning;
  ws_monthly_vaccine_need_details: WsMonthlyVaccineNeedDetails;
// ...
  ws_vaccine_utilization_rate: WsVaccineUtilizationRate;
  ws_village_estimation_details: WsVillageEstimationDetails;
}